+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Bug fix: Correctly handle encoded text as binary, not UTF-8 encoded string
 - Bug fix: External fonts now work like built-in fonts #17
 - Bug fix: Reset colours changed by attributed text
+- Bug fix: Fix global options for text_at/4 when using a string #11
 
 ## 0.3.5
 - Deprecate: `Pdf.delete/1` in favour of `Pdf.cleanup/1`